[refactor] domain모듈 분리
