+++ calculator/src/main/java/edu/nextstep/camp/calculator/CalculatorContract.kt
@@ -1,20 +1,17 @@
 package edu.nextstep.camp.calculator
 
-import edu.nextstep.camp.calculator.domain.Expression
-import edu.nextstep.camp.calculator.domain.Operator
-
 interface CalculatorContract {
     interface View {
         var presenter: Presenter
 
-        fun showExpression(expression: Expression)
+        fun showExpression(expression: edu.nextstep.camp.domain.Expression)
         fun showResult(result: Int)
         fun showIncompleteExpressionError()
     }
 
     interface Presenter {
         fun addToExpression(operand: Int)
-        fun addToExpression(operator: Operator)
+        fun addToExpression(operator: edu.nextstep.camp.domain.Operator)
         fun calculate()
         fun removeLast()
     }
