+++ calculator/src/test/java/edu/nextstep/camp/calculator/domain/ExpressionTest.kt
@@ -7,7 +7,7 @@ class ExpressionTest {
     @Test
     fun `빈 수식일 때, 피연산자를 추가할 수 있어야한다`() {
         // given
-        val expression = Expression.EMPTY
+        val expression = edu.nextstep.camp.domain.Expression.EMPTY
 
         // when
         val actual = expression + 1
@@ -19,7 +19,7 @@ class ExpressionTest {
     @Test
     fun `'8' 수식이 있을 때, 9를 입력하면 89로 바뀌어야 한다`() {
         // given
-        val expression = Expression(listOf(8))
+        val expression = edu.nextstep.camp.domain.Expression(listOf(8))
 
         // when
         val actual = expression + 9
@@ -31,10 +31,10 @@ class ExpressionTest {
     @Test
     fun `빈 수식일 때, + 연산자를 추가할 수 없어야 한다`() {
         // given
-        val expression = Expression.EMPTY
+        val expression = edu.nextstep.camp.domain.Expression.EMPTY
 
         // when
-        val actual = expression + Operator.Plus
+        val actual = expression + edu.nextstep.camp.domain.Operator.Plus
 
         // then
         assertThat(actual.toString()).isEqualTo("")
@@ -43,10 +43,10 @@ class ExpressionTest {
     @Test
     fun `'1' 수식이 있을 때, + 연산자를 추가할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(1))
+        val expression = edu.nextstep.camp.domain.Expression(listOf(1))
 
         // when
-        val actual = expression + Operator.Plus
+        val actual = expression + edu.nextstep.camp.domain.Operator.Plus
 
         // then
         assertThat(actual.toString()).isEqualTo("1 +")
@@ -55,10 +55,11 @@ class ExpressionTest {
     @Test
     fun `'8 +' 수식이 있을 때, + 연산자를 - 연산자로 변경할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(8, Operator.Plus))
+        val expression =
+            edu.nextstep.camp.domain.Expression(listOf(8, edu.nextstep.camp.domain.Operator.Plus))
 
         // when
-        val actual = expression + Operator.Minus
+        val actual = expression + edu.nextstep.camp.domain.Operator.Minus
 
         // then
         assertThat(actual.toString()).isEqualTo("8 -")
@@ -67,7 +68,13 @@ class ExpressionTest {
     @Test
     fun `'32 + 1' 수식이 있을 때, 마지막 1을 제거할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(32, Operator.Plus, 1))
+        val expression = edu.nextstep.camp.domain.Expression(
+            listOf(
+                32,
+                edu.nextstep.camp.domain.Operator.Plus,
+                1
+            )
+        )
 
         // when
         val actual = expression.removeLast()
@@ -79,7 +86,8 @@ class ExpressionTest {
     @Test
     fun `'32 +' 수식이 있을 때, 마지막 +를 제거할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(32, Operator.Plus))
+        val expression =
+            edu.nextstep.camp.domain.Expression(listOf(32, edu.nextstep.camp.domain.Operator.Plus))
 
         // when
         val actual = expression.removeLast()
@@ -91,7 +99,7 @@ class ExpressionTest {
     @Test
     fun `'32' 수식이 있을 때, 마지막 2를 제거할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(32))
+        val expression = edu.nextstep.camp.domain.Expression(listOf(32))
 
         // when
         val actual = expression.removeLast()
@@ -103,7 +111,7 @@ class ExpressionTest {
     @Test
     fun `'3' 수식이 있을 때, 마지막 3을 제거할 수 있어야 한다`() {
         // given
-        val expression = Expression(listOf(3))
+        val expression = edu.nextstep.camp.domain.Expression(listOf(3))
 
         // when
         val actual = expression.removeLast()
@@ -115,7 +123,7 @@ class ExpressionTest {
     @Test
     fun `빈 수식일 때, 마지막을 제거해도 빈 수식이어야 한다`() {
         // given
-        val expression = Expression.EMPTY
+        val expression = edu.nextstep.camp.domain.Expression.EMPTY
 
         // when
         val actual = expression.removeLast()
