+++ domain/src/main/java/edu/nextstep/camp/domain/Expression.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.domain
 
 data class Expression(
     private val values: List<Any> = emptyList()
