+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':calculator'
 include ':counter'
 rootProject.name = "AndroidCalculator"
+include ':domain'
